refactor: DTO 폴더 패키지 정리
