--- backend/src/main/java/com/bookbook/domain/user/dto/UserLoginRequestDto.kt
@@ -1,6 +1,0 @@
-package com.bookbook.domain.user.dto
-
-data class UserLoginRequestDto(
-    var username: String? = null,
-    var password: String? = null
-)
