--- backend/src/main/java/com/bookbook/domain/user/dto/UserStatusResponseDto.kt
@@ -1,8 +1,0 @@
-package com.bookbook.domain.user.dto
-
-import com.bookbook.domain.user.enums.UserStatus
-
-data class UserStatusResponseDto(
-    var id: Long? = null,
-    var userStatus: UserStatus? = null
-)
