+++ backend/src/main/java/com/bookbook/domain/user/dto/request/UserCreateRequestDto.kt
@@ -1,10 +1,10 @@
-package com.bookbook.domain.user.dto
+package com.bookbook.domain.user.dto.request
 
 import jakarta.validation.constraints.NotBlank
 
 data class UserCreateRequestDto(
     @field:NotBlank(message = "아이디는 필수 입력값입니다.")
-    var nickname: String?,
+    var nickname: String,
     @field:NotBlank(message = "주소는 필수 입력값입니다.")
-    var address: String?
+    var address: String
 )
