+++ backend/src/main/java/com/bookbook/domain/user/dto/request/UserLoginRequestDto.kt
@@ -1,0 +1,8 @@
+package com.bookbook.domain.user.dto.request
+
+import org.hibernate.validator.constraints.Length
+
+data class UserLoginRequestDto(
+    @field:Length(min = 2, max = 30) var username: String,
+    @field:Length(min = 2, max = 30) var password: String
+)
