+++ backend/src/main/java/com/bookbook/domain/user/dto/request/UserUpdateRequestDto.kt
@@ -1,4 +1,4 @@
-package com.bookbook.domain.user.dto
+package com.bookbook.domain.user.dto.request
 
 data class UserUpdateRequestDto(
     var nickname: String?,
