+++ backend/src/main/java/com/bookbook/domain/user/dto/response/UserStatusResponseDto.kt
@@ -1,0 +1,8 @@
+package com.bookbook.domain.user.dto.response
+
+import com.bookbook.domain.user.enums.UserStatus
+
+data class UserStatusResponseDto(
+    var id: Long,
+    var userStatus: UserStatus
+)
